// src/utils/enum/acl-enum.ts
export enum ACTIONS {
  PURCHASE_PN_EXPORT = 'purchase.pn.export',
  PURCHASE_PI_CLOSE = 'purchase.pi.close',
}

export enum MAINMENU {
  SALE = 'sale',
  STOCK = 'stock',
}

export enum SUBMENU {
  SALE_DISCOUNT = 'sale.discount',
  STOCK_CHECK = 'stock.checking',
  STOCK_SUPPLY = 'stock.supply',
  STOCK_REQUEST = 'stock.request',
  STOCK_TRANSFER = 'stock.transfer',
}